[MODFISTO-351] Some refactoring
