+++ src/main/java/org/folio/dao/budget/BudgetExpenseClassDAO.java
@@ -1,12 +1,11 @@
 package org.folio.dao.budget;
 
 import io.vertx.core.Future;
-import io.vertx.sqlclient.Tuple;
 import org.folio.rest.jaxrs.model.BudgetExpenseClass;
 import org.folio.rest.persist.DBClient;
 
 import java.util.List;
 
 public interface BudgetExpenseClassDAO {
-  Future<List<BudgetExpenseClass>> getBudgetExpenseClasses(String sql, Tuple params, DBClient client);
+  Future<List<BudgetExpenseClass>> getTemporaryBudgetExpenseClasses(String budgetId, DBClient client);
 }
